Circular Buffer and Interrupt based implementation
This library now uses a simple circular buffer for RX and TX. The user
does not have to fiddle with this data structure as there are methods
that return the data in the buffer. The Buffer is also used to send
data. This way writing to the UART does not block the application. As
soon as new data is in the buffer it will be sent using an interrupt
handler. Receiving data is also interrupt driven.

Doxygen documentation added

The tester has been reworked